requirements.txt
# Django Framework
asgiref==3.8.1
channels==4.2.0
Django==5.1.4
djangorestframework==3.15.2
whitenoise==6.8.2

# Computer Vision
dlib==19.24.6
face-recognition==1.3.0
face-recognition-models==0.3.0
opencv-python==4.10.0.84

# Data Processing
pandas==2.2.3
numpy==2.2.0

# Visualization
matplotlib==3.9.3
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.2
kiwisolver==1.4.7
pillow==11.0.0

# Utilities
click==8.1.7
colorama==0.4.6
packaging==24.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
sqlparse==0.5.2
tzdata==2024.2

# Machine Learning (Optional - for EviTrain module)
# Install separately: cd EviTrain && pip install -r requirements.txt
# catboost>=1.2.8
# lightgbm>=4.6.0
# scikit-learn>=1.7.2
# streamlit>=1.51.0
# seaborn>=0.13.2
# plotly>=6.1.2
# scipy>=1.15.3
# joblib>=1.5.2

# Energy Calculator Dependencies
# pyyaml>=6.0
